requirements.txt
faiss-cpu
sentence-transformers
scikit-learn
numpy
pandas
groq